Move Cluster Operation to apiv1 and add cluster conditions.
