+++ go/admin/v1/adminv1connect/cluster.connect.go
@@ -40,8 +40,6 @@ const (
 	// ClusterServiceCredentialsProcedure is the fully-qualified name of the ClusterService's
 	// Credentials RPC.
 	ClusterServiceCredentialsProcedure = "/admin.v1.ClusterService/Credentials"
-	// ClusterServiceOperateProcedure is the fully-qualified name of the ClusterService's Operate RPC.
-	ClusterServiceOperateProcedure = "/admin.v1.ClusterService/Operate"
 )
 
 // These variables are the protoreflect.Descriptor objects for the RPCs defined in this package.
@@ -50,7 +48,6 @@ var (
 	clusterServiceGetMethodDescriptor         = clusterServiceServiceDescriptor.Methods().ByName("Get")
 	clusterServiceListMethodDescriptor        = clusterServiceServiceDescriptor.Methods().ByName("List")
 	clusterServiceCredentialsMethodDescriptor = clusterServiceServiceDescriptor.Methods().ByName("Credentials")
-	clusterServiceOperateMethodDescriptor     = clusterServiceServiceDescriptor.Methods().ByName("Operate")
 )
 
 // ClusterServiceClient is a client for the admin.v1.ClusterService service.
@@ -61,8 +58,6 @@ type ClusterServiceClient interface {
 	List(context.Context, *connect.Request[v1.ClusterServiceListRequest]) (*connect.Response[v1.ClusterServiceListResponse], error)
 	// Credentials of a cluster
 	Credentials(context.Context, *connect.Request[v1.ClusterServiceCredentialsRequest]) (*connect.Response[v1.ClusterServiceCredentialsResponse], error)
-	// Operate on a cluster
-	Operate(context.Context, *connect.Request[v1.ClusterServiceOperateRequest]) (*connect.Response[v1.ClusterServiceOperateResponse], error)
 }
 
 // NewClusterServiceClient constructs a client for the admin.v1.ClusterService service. By default,
@@ -93,12 +88,6 @@ func NewClusterServiceClient(httpClient connect.HTTPClient, baseURL string, opts
 			connect.WithSchema(clusterServiceCredentialsMethodDescriptor),
 			connect.WithClientOptions(opts...),
 		),
-		operate: connect.NewClient[v1.ClusterServiceOperateRequest, v1.ClusterServiceOperateResponse](
-			httpClient,
-			baseURL+ClusterServiceOperateProcedure,
-			connect.WithSchema(clusterServiceOperateMethodDescriptor),
-			connect.WithClientOptions(opts...),
-		),
 	}
 }
 
@@ -107,7 +96,6 @@ type clusterServiceClient struct {
 	get         *connect.Client[v1.ClusterServiceGetRequest, v1.ClusterServiceGetResponse]
 	list        *connect.Client[v1.ClusterServiceListRequest, v1.ClusterServiceListResponse]
 	credentials *connect.Client[v1.ClusterServiceCredentialsRequest, v1.ClusterServiceCredentialsResponse]
-	operate     *connect.Client[v1.ClusterServiceOperateRequest, v1.ClusterServiceOperateResponse]
 }
 
 // Get calls admin.v1.ClusterService.Get.
@@ -125,11 +113,6 @@ func (c *clusterServiceClient) Credentials(ctx context.Context, req *connect.Req
 	return c.credentials.CallUnary(ctx, req)
 }
 
-// Operate calls admin.v1.ClusterService.Operate.
-func (c *clusterServiceClient) Operate(ctx context.Context, req *connect.Request[v1.ClusterServiceOperateRequest]) (*connect.Response[v1.ClusterServiceOperateResponse], error) {
-	return c.operate.CallUnary(ctx, req)
-}
-
 // ClusterServiceHandler is an implementation of the admin.v1.ClusterService service.
 type ClusterServiceHandler interface {
 	// Get a cluster
@@ -138,8 +121,6 @@ type ClusterServiceHandler interface {
 	List(context.Context, *connect.Request[v1.ClusterServiceListRequest]) (*connect.Response[v1.ClusterServiceListResponse], error)
 	// Credentials of a cluster
 	Credentials(context.Context, *connect.Request[v1.ClusterServiceCredentialsRequest]) (*connect.Response[v1.ClusterServiceCredentialsResponse], error)
-	// Operate on a cluster
-	Operate(context.Context, *connect.Request[v1.ClusterServiceOperateRequest]) (*connect.Response[v1.ClusterServiceOperateResponse], error)
 }
 
 // NewClusterServiceHandler builds an HTTP handler from the service implementation. It returns the
@@ -166,12 +147,6 @@ func NewClusterServiceHandler(svc ClusterServiceHandler, opts ...connect.Handler
 		connect.WithSchema(clusterServiceCredentialsMethodDescriptor),
 		connect.WithHandlerOptions(opts...),
 	)
-	clusterServiceOperateHandler := connect.NewUnaryHandler(
-		ClusterServiceOperateProcedure,
-		svc.Operate,
-		connect.WithSchema(clusterServiceOperateMethodDescriptor),
-		connect.WithHandlerOptions(opts...),
-	)
 	return "/admin.v1.ClusterService/", http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
 		switch r.URL.Path {
 		case ClusterServiceGetProcedure:
@@ -180,8 +155,6 @@ func NewClusterServiceHandler(svc ClusterServiceHandler, opts ...connect.Handler
 			clusterServiceListHandler.ServeHTTP(w, r)
 		case ClusterServiceCredentialsProcedure:
 			clusterServiceCredentialsHandler.ServeHTTP(w, r)
-		case ClusterServiceOperateProcedure:
-			clusterServiceOperateHandler.ServeHTTP(w, r)
 		default:
 			http.NotFound(w, r)
 		}
@@ -201,8 +174,4 @@ func (UnimplementedClusterServiceHandler) List(context.Context, *connect.Request
 
 func (UnimplementedClusterServiceHandler) Credentials(context.Context, *connect.Request[v1.ClusterServiceCredentialsRequest]) (*connect.Response[v1.ClusterServiceCredentialsResponse], error) {
 	return nil, connect.NewError(connect.CodeUnimplemented, errors.New("admin.v1.ClusterService.Credentials is not implemented"))
-}
-
-func (UnimplementedClusterServiceHandler) Operate(context.Context, *connect.Request[v1.ClusterServiceOperateRequest]) (*connect.Response[v1.ClusterServiceOperateResponse], error) {
-	return nil, connect.NewError(connect.CodeUnimplemented, errors.New("admin.v1.ClusterService.Operate is not implemented"))
 }
