+++ go/api/v1/apiv1connect/cluster.connect.go
@@ -49,6 +49,8 @@ const (
 	// ClusterServiceGetCredentialsProcedure is the fully-qualified name of the ClusterService's
 	// GetCredentials RPC.
 	ClusterServiceGetCredentialsProcedure = "/api.v1.ClusterService/GetCredentials"
+	// ClusterServiceOperateProcedure is the fully-qualified name of the ClusterService's Operate RPC.
+	ClusterServiceOperateProcedure = "/api.v1.ClusterService/Operate"
 )
 
 // These variables are the protoreflect.Descriptor objects for the RPCs defined in this package.
@@ -61,6 +63,7 @@ var (
 	clusterServiceDeleteMethodDescriptor         = clusterServiceServiceDescriptor.Methods().ByName("Delete")
 	clusterServiceUpdateMethodDescriptor         = clusterServiceServiceDescriptor.Methods().ByName("Update")
 	clusterServiceGetCredentialsMethodDescriptor = clusterServiceServiceDescriptor.Methods().ByName("GetCredentials")
+	clusterServiceOperateMethodDescriptor        = clusterServiceServiceDescriptor.Methods().ByName("Operate")
 )
 
 // ClusterServiceClient is a client for the api.v1.ClusterService service.
@@ -79,6 +82,8 @@ type ClusterServiceClient interface {
 	Update(context.Context, *connect.Request[v1.ClusterServiceUpdateRequest]) (*connect.Response[v1.ClusterServiceUpdateResponse], error)
 	// GetCredentials of a cluster
 	GetCredentials(context.Context, *connect.Request[v1.ClusterServiceGetCredentialsRequest]) (*connect.Response[v1.ClusterServiceGetCredentialsResponse], error)
+	// Operate on a cluster
+	Operate(context.Context, *connect.Request[v1.ClusterServiceOperateRequest]) (*connect.Response[v1.ClusterServiceOperateResponse], error)
 }
 
 // NewClusterServiceClient constructs a client for the api.v1.ClusterService service. By default, it
@@ -133,6 +138,12 @@ func NewClusterServiceClient(httpClient connect.HTTPClient, baseURL string, opts
 			connect.WithSchema(clusterServiceGetCredentialsMethodDescriptor),
 			connect.WithClientOptions(opts...),
 		),
+		operate: connect.NewClient[v1.ClusterServiceOperateRequest, v1.ClusterServiceOperateResponse](
+			httpClient,
+			baseURL+ClusterServiceOperateProcedure,
+			connect.WithSchema(clusterServiceOperateMethodDescriptor),
+			connect.WithClientOptions(opts...),
+		),
 	}
 }
 
@@ -145,6 +156,7 @@ type clusterServiceClient struct {
 	delete         *connect.Client[v1.ClusterServiceDeleteRequest, v1.ClusterServiceDeleteResponse]
 	update         *connect.Client[v1.ClusterServiceUpdateRequest, v1.ClusterServiceUpdateResponse]
 	getCredentials *connect.Client[v1.ClusterServiceGetCredentialsRequest, v1.ClusterServiceGetCredentialsResponse]
+	operate        *connect.Client[v1.ClusterServiceOperateRequest, v1.ClusterServiceOperateResponse]
 }
 
 // Create calls api.v1.ClusterService.Create.
@@ -182,6 +194,11 @@ func (c *clusterServiceClient) GetCredentials(ctx context.Context, req *connect.
 	return c.getCredentials.CallUnary(ctx, req)
 }
 
+// Operate calls api.v1.ClusterService.Operate.
+func (c *clusterServiceClient) Operate(ctx context.Context, req *connect.Request[v1.ClusterServiceOperateRequest]) (*connect.Response[v1.ClusterServiceOperateResponse], error) {
+	return c.operate.CallUnary(ctx, req)
+}
+
 // ClusterServiceHandler is an implementation of the api.v1.ClusterService service.
 type ClusterServiceHandler interface {
 	// Create a cluster
@@ -198,6 +215,8 @@ type ClusterServiceHandler interface {
 	Update(context.Context, *connect.Request[v1.ClusterServiceUpdateRequest]) (*connect.Response[v1.ClusterServiceUpdateResponse], error)
 	// GetCredentials of a cluster
 	GetCredentials(context.Context, *connect.Request[v1.ClusterServiceGetCredentialsRequest]) (*connect.Response[v1.ClusterServiceGetCredentialsResponse], error)
+	// Operate on a cluster
+	Operate(context.Context, *connect.Request[v1.ClusterServiceOperateRequest]) (*connect.Response[v1.ClusterServiceOperateResponse], error)
 }
 
 // NewClusterServiceHandler builds an HTTP handler from the service implementation. It returns the
@@ -248,6 +267,12 @@ func NewClusterServiceHandler(svc ClusterServiceHandler, opts ...connect.Handler
 		connect.WithSchema(clusterServiceGetCredentialsMethodDescriptor),
 		connect.WithHandlerOptions(opts...),
 	)
+	clusterServiceOperateHandler := connect.NewUnaryHandler(
+		ClusterServiceOperateProcedure,
+		svc.Operate,
+		connect.WithSchema(clusterServiceOperateMethodDescriptor),
+		connect.WithHandlerOptions(opts...),
+	)
 	return "/api.v1.ClusterService/", http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
 		switch r.URL.Path {
 		case ClusterServiceCreateProcedure:
@@ -264,6 +289,8 @@ func NewClusterServiceHandler(svc ClusterServiceHandler, opts ...connect.Handler
 			clusterServiceUpdateHandler.ServeHTTP(w, r)
 		case ClusterServiceGetCredentialsProcedure:
 			clusterServiceGetCredentialsHandler.ServeHTTP(w, r)
+		case ClusterServiceOperateProcedure:
+			clusterServiceOperateHandler.ServeHTTP(w, r)
 		default:
 			http.NotFound(w, r)
 		}
@@ -299,4 +326,8 @@ func (UnimplementedClusterServiceHandler) Update(context.Context, *connect.Reque
 
 func (UnimplementedClusterServiceHandler) GetCredentials(context.Context, *connect.Request[v1.ClusterServiceGetCredentialsRequest]) (*connect.Response[v1.ClusterServiceGetCredentialsResponse], error) {
 	return nil, connect.NewError(connect.CodeUnimplemented, errors.New("api.v1.ClusterService.GetCredentials is not implemented"))
+}
+
+func (UnimplementedClusterServiceHandler) Operate(context.Context, *connect.Request[v1.ClusterServiceOperateRequest]) (*connect.Response[v1.ClusterServiceOperateResponse], error) {
+	return nil, connect.NewError(connect.CodeUnimplemented, errors.New("api.v1.ClusterService.Operate is not implemented"))
 }
