+++ go/permissions/servicepermissions.go
@@ -33,7 +33,6 @@ func GetServicePermissions() *ServicePermissions {
 					"/admin.v1.ClusterService/Get",
 					"/admin.v1.ClusterService/List",
 					"/admin.v1.ClusterService/Credentials",
-					"/admin.v1.ClusterService/Operate",
 					"/admin.v1.PaymentService/ListCoupons",
 					"/admin.v1.PaymentService/AddCouponToCustomer",
 					"/admin.v1.PaymentService/AddBalanceToCustomer",
@@ -45,6 +44,7 @@ func GetServicePermissions() *ServicePermissions {
 					"/admin.v1.TenantService/Revoke",
 					"/admin.v1.TokenService/List",
 					"/admin.v1.TokenService/Revoke",
+					"/api.v1.ClusterService/Operate",
 				},
 				Viewer: []string{
 					"/admin.v1.ClusterService/Get",
@@ -154,7 +154,6 @@ func GetServicePermissions() *ServicePermissions {
 			"/admin.v1.ClusterService/Credentials":          true,
 			"/admin.v1.ClusterService/Get":                  true,
 			"/admin.v1.ClusterService/List":                 true,
-			"/admin.v1.ClusterService/Operate":              true,
 			"/admin.v1.PaymentService/AddBalanceToCustomer": true,
 			"/admin.v1.PaymentService/AddCouponToCustomer":  true,
 			"/admin.v1.PaymentService/ListCoupons":          true,
@@ -172,6 +171,7 @@ func GetServicePermissions() *ServicePermissions {
 			"/api.v1.ClusterService/Get":                    true,
 			"/api.v1.ClusterService/GetCredentials":         true,
 			"/api.v1.ClusterService/List":                   true,
+			"/api.v1.ClusterService/Operate":                true,
 			"/api.v1.ClusterService/Update":                 true,
 			"/api.v1.ClusterService/WatchStatus":            true,
 			"/api.v1.HealthService/Get":                     true,
@@ -247,7 +247,6 @@ func GetServicePermissions() *ServicePermissions {
 			"/admin.v1.ClusterService/Credentials":          true,
 			"/admin.v1.ClusterService/Get":                  true,
 			"/admin.v1.ClusterService/List":                 true,
-			"/admin.v1.ClusterService/Operate":              true,
 			"/admin.v1.PaymentService/AddBalanceToCustomer": true,
 			"/admin.v1.PaymentService/AddCouponToCustomer":  true,
 			"/admin.v1.PaymentService/ListCoupons":          true,
@@ -265,6 +264,7 @@ func GetServicePermissions() *ServicePermissions {
 			"/api.v1.ClusterService/Get":                    false,
 			"/api.v1.ClusterService/GetCredentials":         true,
 			"/api.v1.ClusterService/List":                   false,
+			"/api.v1.ClusterService/Operate":                true,
 			"/api.v1.ClusterService/Update":                 true,
 			"/api.v1.ClusterService/WatchStatus":            false,
 			"/api.v1.HealthService/Get":                     false,
