+++ go/tests/mocks/admin/v1/adminv1connect/ClusterServiceClient.go
@@ -106,36 +106,6 @@ func (_m *ClusterServiceClient) List(_a0 context.Context, _a1 *connect.Request[a
 	return r0, r1
 }
 
-// Operate provides a mock function with given fields: _a0, _a1
-func (_m *ClusterServiceClient) Operate(_a0 context.Context, _a1 *connect.Request[adminv1.ClusterServiceOperateRequest]) (*connect.Response[adminv1.ClusterServiceOperateResponse], error) {
-	ret := _m.Called(_a0, _a1)
-
-	if len(ret) == 0 {
-		panic("no return value specified for Operate")
-	}
-
-	var r0 *connect.Response[adminv1.ClusterServiceOperateResponse]
-	var r1 error
-	if rf, ok := ret.Get(0).(func(context.Context, *connect.Request[adminv1.ClusterServiceOperateRequest]) (*connect.Response[adminv1.ClusterServiceOperateResponse], error)); ok {
-		return rf(_a0, _a1)
-	}
-	if rf, ok := ret.Get(0).(func(context.Context, *connect.Request[adminv1.ClusterServiceOperateRequest]) *connect.Response[adminv1.ClusterServiceOperateResponse]); ok {
-		r0 = rf(_a0, _a1)
-	} else {
-		if ret.Get(0) != nil {
-			r0 = ret.Get(0).(*connect.Response[adminv1.ClusterServiceOperateResponse])
-		}
-	}
-
-	if rf, ok := ret.Get(1).(func(context.Context, *connect.Request[adminv1.ClusterServiceOperateRequest]) error); ok {
-		r1 = rf(_a0, _a1)
-	} else {
-		r1 = ret.Error(1)
-	}
-
-	return r0, r1
-}
-
 // NewClusterServiceClient creates a new instance of ClusterServiceClient. It also registers a testing interface on the mock and a cleanup function to assert the mocks expectations.
 // The first argument is typically a *testing.T value.
 func NewClusterServiceClient(t interface {
