+++ go/tests/mocks/api/v1/apiv1connect/ClusterServiceHandler.go
@@ -166,6 +166,36 @@ func (_m *ClusterServiceHandler) List(_a0 context.Context, _a1 *connect.Request[
 	return r0, r1
 }
 
+// Operate provides a mock function with given fields: _a0, _a1
+func (_m *ClusterServiceHandler) Operate(_a0 context.Context, _a1 *connect.Request[apiv1.ClusterServiceOperateRequest]) (*connect.Response[apiv1.ClusterServiceOperateResponse], error) {
+	ret := _m.Called(_a0, _a1)
+
+	if len(ret) == 0 {
+		panic("no return value specified for Operate")
+	}
+
+	var r0 *connect.Response[apiv1.ClusterServiceOperateResponse]
+	var r1 error
+	if rf, ok := ret.Get(0).(func(context.Context, *connect.Request[apiv1.ClusterServiceOperateRequest]) (*connect.Response[apiv1.ClusterServiceOperateResponse], error)); ok {
+		return rf(_a0, _a1)
+	}
+	if rf, ok := ret.Get(0).(func(context.Context, *connect.Request[apiv1.ClusterServiceOperateRequest]) *connect.Response[apiv1.ClusterServiceOperateResponse]); ok {
+		r0 = rf(_a0, _a1)
+	} else {
+		if ret.Get(0) != nil {
+			r0 = ret.Get(0).(*connect.Response[apiv1.ClusterServiceOperateResponse])
+		}
+	}
+
+	if rf, ok := ret.Get(1).(func(context.Context, *connect.Request[apiv1.ClusterServiceOperateRequest]) error); ok {
+		r1 = rf(_a0, _a1)
+	} else {
+		r1 = ret.Error(1)
+	}
+
+	return r0, r1
+}
+
 // Update provides a mock function with given fields: _a0, _a1
 func (_m *ClusterServiceHandler) Update(_a0 context.Context, _a1 *connect.Request[apiv1.ClusterServiceUpdateRequest]) (*connect.Response[apiv1.ClusterServiceUpdateResponse], error) {
 	ret := _m.Called(_a0, _a1)
