+++ js/admin/v1/cluster_connect.ts
@@ -3,7 +3,7 @@
 /* eslint-disable */
 // @ts-nocheck
 
-import { ClusterServiceCredentialsRequest, ClusterServiceCredentialsResponse, ClusterServiceGetRequest, ClusterServiceGetResponse, ClusterServiceListRequest, ClusterServiceListResponse, ClusterServiceOperateRequest, ClusterServiceOperateResponse } from "./cluster_pb.js";
+import { ClusterServiceCredentialsRequest, ClusterServiceCredentialsResponse, ClusterServiceGetRequest, ClusterServiceGetResponse, ClusterServiceListRequest, ClusterServiceListResponse } from "./cluster_pb.js";
 import { MethodKind } from "@bufbuild/protobuf";
 
 /**
@@ -45,17 +45,6 @@ export const ClusterService = {
       name: "Credentials",
       I: ClusterServiceCredentialsRequest,
       O: ClusterServiceCredentialsResponse,
-      kind: MethodKind.Unary,
-    },
-    /**
-     * Operate on a cluster
-     *
-     * @generated from rpc admin.v1.ClusterService.Operate
-     */
-    operate: {
-      name: "Operate",
-      I: ClusterServiceOperateRequest,
-      O: ClusterServiceOperateResponse,
       kind: MethodKind.Unary,
     },
   }
