+++ js/admin/v1/cluster_pb.d.ts
@@ -3,37 +3,6 @@ import { Duration, Message, proto3 } from "@bufbuild/protobuf";
 import { Cluster } from "../../api/v1/cluster_pb.js";
 import { Machine, VPN } from "./machine_pb.js";
 /**
- * Operate defines the types of reconcilation to be triggered
- *
- * @generated from enum admin.v1.Operate
- */
-export declare enum Operate {
-    /**
-     * OPERATE_UNSPECIFIED is not specified
-     *
-     * @generated from enum value: OPERATE_UNSPECIFIED = 0;
-     */
-    UNSPECIFIED = 0,
-    /**
-     * OPERATE_RECONCILE reconcile the cluster
-     *
-     * @generated from enum value: OPERATE_RECONCILE = 1;
-     */
-    RECONCILE = 1,
-    /**
-     * OPERATE_MAINTAIN maintain the cluster
-     *
-     * @generated from enum value: OPERATE_MAINTAIN = 2;
-     */
-    MAINTAIN = 2,
-    /**
-     * OPERATE_RETRY retry the reconcilation of the cluster
-     *
-     * @generated from enum value: OPERATE_RETRY = 3;
-     */
-    RETRY = 3
-}
-/**
  * ClusterServiceGetRequest is the request payload for the cluster get request
  *
  * @generated from message admin.v1.ClusterServiceGetRequest
@@ -165,33 +134,6 @@ export declare class ClusterServiceCredentialsRequest extends Message<ClusterSer
     static equals(a: ClusterServiceCredentialsRequest | PlainMessage<ClusterServiceCredentialsRequest> | undefined, b: ClusterServiceCredentialsRequest | PlainMessage<ClusterServiceCredentialsRequest> | undefined): boolean;
 }
 /**
- * ClusterServiceOperateRequest is the request payload for the cluster operate request
- *
- * @generated from message admin.v1.ClusterServiceOperateRequest
- */
-export declare class ClusterServiceOperateRequest extends Message<ClusterServiceOperateRequest> {
-    /**
-     * Uuid of the cluster
-     *
-     * @generated from field: string uuid = 1;
-     */
-    uuid: string;
-    /**
-     * Operate is the reconcilation operation which should be performed
-     *
-     * @generated from field: admin.v1.Operate operate = 2;
-     */
-    operate: Operate;
-    constructor(data?: PartialMessage<ClusterServiceOperateRequest>);
-    static readonly runtime: typeof proto3;
-    static readonly typeName = "admin.v1.ClusterServiceOperateRequest";
-    static readonly fields: FieldList;
-    static fromBinary(bytes: Uint8Array, options?: Partial<BinaryReadOptions>): ClusterServiceOperateRequest;
-    static fromJson(jsonValue: JsonValue, options?: Partial<JsonReadOptions>): ClusterServiceOperateRequest;
-    static fromJsonString(jsonString: string, options?: Partial<JsonReadOptions>): ClusterServiceOperateRequest;
-    static equals(a: ClusterServiceOperateRequest | PlainMessage<ClusterServiceOperateRequest> | undefined, b: ClusterServiceOperateRequest | PlainMessage<ClusterServiceOperateRequest> | undefined): boolean;
-}
-/**
  * ClusterServiceGetResponse is the response payload for the cluster get request
  *
  * @generated from message admin.v1.ClusterServiceGetResponse
@@ -271,27 +213,6 @@ export declare class ClusterServiceCredentialsResponse extends Message<ClusterSe
     static fromJson(jsonValue: JsonValue, options?: Partial<JsonReadOptions>): ClusterServiceCredentialsResponse;
     static fromJsonString(jsonString: string, options?: Partial<JsonReadOptions>): ClusterServiceCredentialsResponse;
     static equals(a: ClusterServiceCredentialsResponse | PlainMessage<ClusterServiceCredentialsResponse> | undefined, b: ClusterServiceCredentialsResponse | PlainMessage<ClusterServiceCredentialsResponse> | undefined): boolean;
-}
-/**
- * ClusterServiceOperateResponse is the response payload for the cluster operate request
- *
- * @generated from message admin.v1.ClusterServiceOperateResponse
- */
-export declare class ClusterServiceOperateResponse extends Message<ClusterServiceOperateResponse> {
-    /**
-     * Cluster is the cluster
-     *
-     * @generated from field: api.v1.Cluster cluster = 1;
-     */
-    cluster?: Cluster;
-    constructor(data?: PartialMessage<ClusterServiceOperateResponse>);
-    static readonly runtime: typeof proto3;
-    static readonly typeName = "admin.v1.ClusterServiceOperateResponse";
-    static readonly fields: FieldList;
-    static fromBinary(bytes: Uint8Array, options?: Partial<BinaryReadOptions>): ClusterServiceOperateResponse;
-    static fromJson(jsonValue: JsonValue, options?: Partial<JsonReadOptions>): ClusterServiceOperateResponse;
-    static fromJsonString(jsonString: string, options?: Partial<JsonReadOptions>): ClusterServiceOperateResponse;
-    static equals(a: ClusterServiceOperateResponse | PlainMessage<ClusterServiceOperateResponse> | undefined, b: ClusterServiceOperateResponse | PlainMessage<ClusterServiceOperateResponse> | undefined): boolean;
 }
 /**
  * SSHKeyPair details to access a firewall via ssh
