+++ js/permissions/servicepermissions.json
@@ -5,7 +5,6 @@
         "/admin.v1.ClusterService/Get",
         "/admin.v1.ClusterService/List",
         "/admin.v1.ClusterService/Credentials",
-        "/admin.v1.ClusterService/Operate",
         "/admin.v1.PaymentService/ListCoupons",
         "/admin.v1.PaymentService/AddCouponToCustomer",
         "/admin.v1.PaymentService/AddBalanceToCustomer",
@@ -16,7 +15,8 @@
         "/admin.v1.TenantService/Admit",
         "/admin.v1.TenantService/Revoke",
         "/admin.v1.TokenService/List",
-        "/admin.v1.TokenService/Revoke"
+        "/admin.v1.TokenService/Revoke",
+        "/api.v1.ClusterService/Operate"
       ],
       "viewer": [
         "/admin.v1.ClusterService/Get",
@@ -126,7 +126,6 @@
     "/admin.v1.ClusterService/Credentials": true,
     "/admin.v1.ClusterService/Get": true,
     "/admin.v1.ClusterService/List": true,
-    "/admin.v1.ClusterService/Operate": true,
     "/admin.v1.PaymentService/AddBalanceToCustomer": true,
     "/admin.v1.PaymentService/AddCouponToCustomer": true,
     "/admin.v1.PaymentService/ListCoupons": true,
@@ -144,6 +143,7 @@
     "/api.v1.ClusterService/Get": true,
     "/api.v1.ClusterService/GetCredentials": true,
     "/api.v1.ClusterService/List": true,
+    "/api.v1.ClusterService/Operate": true,
     "/api.v1.ClusterService/Update": true,
     "/api.v1.ClusterService/WatchStatus": true,
     "/api.v1.HealthService/Get": true,
@@ -219,7 +219,6 @@
     "/admin.v1.ClusterService/Credentials": true,
     "/admin.v1.ClusterService/Get": true,
     "/admin.v1.ClusterService/List": true,
-    "/admin.v1.ClusterService/Operate": true,
     "/admin.v1.PaymentService/AddBalanceToCustomer": true,
     "/admin.v1.PaymentService/AddCouponToCustomer": true,
     "/admin.v1.PaymentService/ListCoupons": true,
@@ -237,6 +236,7 @@
     "/api.v1.ClusterService/Get": false,
     "/api.v1.ClusterService/GetCredentials": true,
     "/api.v1.ClusterService/List": false,
+    "/api.v1.ClusterService/Operate": true,
     "/api.v1.ClusterService/Update": true,
     "/api.v1.ClusterService/WatchStatus": false,
     "/api.v1.HealthService/Get": false,
